pyproject.toml
[tool.poetry]
name = "fpl_draft"
version = "0.1.0"
description = "project to help decision making in fpl draft"
authors = ["jberm0 <jmberman12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
polars = "^1.4.1"
poetry-core = "^1.8.3"
pre-commit = "^3.8.0"
virtualenv = "^20.26.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
